handle 410 http responses
If an inbound Inbox item fails because `raise_is_blocked_activity` identifies the sender has been deleted or blocked on their server, then it is clearer to send an `HttpResponseForbidden` back to them, rather than ambiguously raising a `PermissionDenied` error implying we did something wrong at our end.

Additionally, when any ActivityPub request returns a `410` (regardless of the object type) we should log this only as a `WARNING` rather than an exception, since it is not really an error that a system admin can do anything about.

This commit changes behaviour in three places:

If there is an `HTTPError`, we now:

* check to see whether the status code is `410`
* log a warning (instead of an exception)
* `delete()` the user locally if they are not already marked as deleted

When identifying that the sending user has been deleted, instead of raising `PermissionDenied()` we raise a new custom error – `UserIsGoneError()`

When calling `raise_is_blocked_activity()` we now use a try..catch block that catches any `UserIsGoneError.` If `UserIsGoneError` is identified, we return `HttpResponseForbidden()` (previously this was `PermissionDenied()` passing through from `raise_is_blocked_activity`)

fixes #3552